Fixed ast test. Deleted program_to_run_AKV, slight changes
