+++ src/.gitignore
@@ -1,3 +1,4 @@
 *.swp
 1
 test_prog_to_run.rs
+alugrammar.rs
